Add tests and move get_comments_above_step_in_dag to helpers
